Native.Signal.PointerLock: corrected syntax errors
It is with a great heaviness of heart that, mere hours after sending the
punch cards to the publication facility for public scrutiny, I discovered
some syntactical errors in my previously cargo-culted implementation of
PointerLock.

Of course, one notes that modern technology allows a developer to actually
compile and run code before considering it ready for any kind of use, but
I have little time for such trifles.